+++ apps/pleiepenger-sykt-barn/src/app/søknad/legeerklæring-step/LegeerklæringStep.tsx
@@ -22,7 +22,7 @@ import getLenker from '../../lenker';
 import { StepCommonProps } from '../../types/StepCommonProps';
 import { StepID } from '../../types/StepID';
 import { SøknadFormField, SøknadFormValues } from '../../types/søknad-form-values/SøknadFormValues';
-import { getAttachmentURLFrontend } from '../../utils/attachmentUtilsAuthToken';
+import { getAttachmentURLFrontend } from '../../utils/appAttachmentUtils';
 import { relocateToLoginPage } from '../../utils/navigationUtils';
 import { validateLegeerklæring } from '../../validation/fieldValidations';
 import SøknadFormStep from '../SøknadFormStep';